Criada as classes e seus repositories para matricula 'Enrollment' e desistencias 'Dropout'

+++ src/main/java/com/dougfsilva/e_AGE/application/usecases/responsible/DeleteResponsible.java
@@ -1,12 +1,11 @@
 package com.dougfsilva.e_AGE.application.usecases.responsible;
 
-import java.util.List;
-
+import com.dougfsilva.e_AGE.application.usecases.address.DeleteAddress;
 import com.dougfsilva.e_AGE.application.usecases.utilities.StandardLogger;
 import com.dougfsilva.e_AGE.domain.responsible.Responsible;
 import com.dougfsilva.e_AGE.domain.responsible.ResponsibleRepository;
-import com.dougfsilva.e_AGE.domain.student.Student;
 import com.dougfsilva.e_AGE.domain.student.StudentRepository;
+import com.dougfsilva.e_AGE.domain.utilities.exception.DataIntegrityViolationException;
 
 import lombok.AllArgsConstructor;
 
@@ -19,16 +18,19 @@ public class DeleteResponsible {
 	
 	private final FindResponsible findResponsible;
 	
+	private final DeleteAddress deleteAddress;
+	
 	private final StandardLogger logger;
 	
 	public void execute(String ID) {
 		Responsible responsible = findResponsible.findByID(ID);
-		List<Student> students = studentRepository.findAllByResponsible(responsible);
-		students.forEach(student -> {
-			student.setResponsible(null);
-			studentRepository.save(student);
-		});
+		if(studentRepository.existsByResponsible(responsible)) {
+			throw new DataIntegrityViolationException(
+					String.format("The Responsible %S cannot be deleted because there are student(s) still associated with it!", 
+							responsible.getName()));
+		}
 		repository.delete(responsible);
+		deleteAddress.execute(responsible.getAddress().getID());
 		logger.deletedObjectLog(responsible);
 	}
 }
